% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cuisine_in_area.R
\name{best_cuisine_in_area}
\alias{best_cuisine_in_area}
\title{Find the Best Cuisine in a Specified Area}
\usage{
best_cuisine_in_area(location = "Vancouver, BC", limit = 10)
}
\arguments{
\item{location}{A string specifying the city (default: "Vancouver, BC").}

\item{limit}{An integer specifying the number of restaurants to retrieve.}
}
\value{
A tibble containing cuisine categories, average ratings, and restaurant counts.
}
\description{
Aggregates restaurant ratings to identify the highest-rated cuisines in a location.
}
