% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restaurant_details.R
\name{get_restaurant_details}
\alias{get_restaurant_details}
\title{Get Restaurant Details by Name or Address}
\usage{
get_restaurant_details(query, location = "Vancouver, BC")
}
\arguments{
\item{query}{A string representing the restaurant name or address.}

\item{location}{A string specifying the city or area (default: "Vancouver, BC").}
}
\value{
A tibble containing restaurant details such as rating, address, phone, and Yelp URL.
}
\description{
Retrieves detailed information about a restaurant based on a name or address search.
}
