% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yelp_api_key.R
\name{get_yelp_api_key}
\alias{get_yelp_api_key}
\title{Get Yelp API Key}
\usage{
get_yelp_api_key()
}
\value{
A string containing the Yelp API key.
}
\description{
Retrieves the Yelp API key from environment variables.
}
