% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_locations.R
\name{list_available_locations}
\alias{list_available_locations}
\title{List Available Locations}
\usage{
list_available_locations()
}
\value{
A tibble containing city names.
}
\description{
Returns a list of predefined cities available for restaurant searches.
}
