% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_pdf.R
\name{open_image}
\alias{open_image}
\title{Open an image or PDF on the desktop}
\usage{
open_image(file)
}
\arguments{
\item{file}{The complete path to an image or PDF file.}
}
\description{
Open an image or PDF on the desktop
}
