% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_url.R
\name{open_url}
\alias{open_url}
\title{Open URL in a desktop web browser}
\usage{
open_url(url, browser = "safari")
}
\arguments{
\item{url}{The URL that we want to view.}

\item{browser}{The name of the web browser.}
}
\description{
Open URL in a desktop web browser
}
